remove unnecessary component between components
